requirements.txt
awscli==1.20.15
beautifulsoup4==4.8.2
GitPython==3.0.8
boto3==1.18.15
Flask==2.1.1
glean_parser==5.0.1
google-cloud-storage==2.2.1
Jinja2==3.1.1
jsonschema==3.1.1
python-dateutil==2.8.0
PyYAML==5.4.1
requests==2.26.0
requests_cache==0.5.2
requests_file==1.4.3
schema==0.7.1
urllib3==1.26.6
Werkzeug==2.1.1
yamllint
